backend/api/requirements.txt
Flask
flask-cors
Flask-Limiter
flask-jwt-extended
xmltodict
requests
python-qbittorrent